requirements.txt
pillow>=10.0.0
exifread>=3.0.0
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.30.0
pandas>=2.0.0
